Some renaming and add IDirectoryCommon
