--- src/models/Directory.ts
@@ -1,6 +1,0 @@
-import { Script } from "./Script";
-
-export class Directory {
-    constructor(public directories: Directory[], public scripts: Script[])
-    {}
-}
