+++ src/models/IDirectory.ts
@@ -1,0 +1,7 @@
+import { IDirectoryCommon } from "./IDirectoryCommon";
+import { ScriptId } from "./ScriptId";
+
+export interface IDirectory {
+    common: IDirectoryCommon
+    _id: ScriptId
+}
