+++ src/models/IDirectoryCommon.ts
@@ -1,0 +1,4 @@
+export interface IDirectoryCommon {
+    name?: string;
+    expert?: boolean;
+}
