+++ src/models/IScriptCommon.ts
